tests: Add an installation test
We need to test the installation process from a live system.

Also, in order to run subsequent tests from a qcow2 image, we actually
need to install into that qcow2 image so we have an installed disk to
boot from for the subsequent tests.

Run the live boot test separately.

This is a bit of a monster of a commit, adding a separate FBE test path
for the process of installing using the FBE wizard. It does the
installation, then reboots into the new system and runs through the FBE
user creation.

Signed-off-by: Philip Withnall <withnall@endlessm.com>