tests: Revert workaround for KVM not being enabled on worker machine
We are now on a packet.net machine which supports KVM, so enable that
for the x86_64 worker to speed things up 100-fold.

This is a modified revert of 80c67bf52765250f220d0e418d822f8ca2ca15e9,
since some of the tests have been split up in the meantime.

Signed-off-by: Philip Withnall <withnall@endlessm.com>