tests: Support OS updates at the start of a test run
In order to test that installations updated from an older version of
the OS continue to work the same as a new installation, we need to
actually be able to update the OS. Do that as a step immediately after
installation, triggered by the presence of an OS_UPDATE_TO variable in
the test dictionary.

Document it all.

A companion change will need to be made to eos-image-builder to get it
to trigger tests with OS_UPDATE_TO set.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T23693