lib: Dynamically inject fake curl implementation
The built-in upload_logs() implementation uses curl to submit form data
to the test harness, containing the logs to upload. EOS doesn’t have
curl installed by default, and we can’t install it by default in the
images just to satisfy OpenQA — that would pointlessly increase the
image size and wouldn’t work for the older releases which we have to
test.

So, we now have a fake curl implementation made using the excellent
Python requests module. lib/installedtest.pm uses the classic ‘wget,
copy to $PATH and chmod +x’ technique to inject it into $PATH, in a
spectacular display of all kinds of antipattern. Kids, do not do this at
home.

This is all triggered by the new ensure_curl_available() method, which
needs to be called before anyone tries using upload_logs().

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T23693