tests: Enable coredumps after installation
It turns out that we don’t get crashes listed in the journal at all
unless coredumpd is enabled fully. I was expecting there to be a
message in the journal about each crash, but not a coredump file or
(necessarily) a listing in coredumpctl.

Fix that by running eos-enable-coredumps after installation, before the
qcow2 image is saved, so it persists in the images used for the rest of
the tests.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T23693