templates: Update to use the image type in the FLAVOR variable
The image builder is being changed to include the image type (‘full’,
‘boot’ or ‘iso’) and optionally ‘_update’ in the FLAVOR variable, so
that test runs of the same image build with different media don’t
overwrite each other’s results.

Update the templates file to take this into account, duplicating most
tests so they’re run for raw images and ISOs. Notable exceptions are the
live boot tests (only runnable on ISOs) and the eosinstaller and
eosnonfree products, which we don’t build ISOs for.

Nothing has been done about update tests in this commit. They’ll be
covered in a following commit.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T23709