tests: Always skip Chrome ‘default browser’ dialogue
I previously thought that the ‘default browser’ dialogue would only
appear when Chrome was run for the first time after installation; but it
appears that there are cases where it can be installed in the background
but not run before this test tries to run it — in which case, the
‘default browser’ dialogue will be displayed. Make the test skip it
unconditionally. If the dialogue isn’t shown, pressing enter at that
point should hopefully be harmless anyway.

Signed-off-by: Philip Withnall <withnall@endlessm.com>