eos-openqa-client: Use images.endlessm-sf.com for images host
This is the new canonical hostname. The public images.endlessm.com name
will be dropped when there are no more external users. The server will
only be accessible from the internal network after that. This should be
fine for the OpenQA server as it's now connected to the internal network
and can resolve endlessm-sf.com internal domains.

https://phabricator.endlessm.com/T23750