tests: Handle installing Chrome on first startup of gnome-software
When running gnome-software for the first time on a newly installed
system, it will automatically install Google Chrome. We want to ignore
that so we can get back to the main screen and install VLC.

We can’t just wait for it to finish, since it needs some UI interaction
once Chrome’s installed to get back to the gnome-software main screen.
Implement that by quitting and restarting gnome-software — it shouldn’t
try and install Chrome a second time.

Signed-off-by: Philip Withnall <withnall@endlessm.com>