eibopenqaserver: send request data as body
It is common for HTTP servers to reject requests whose URLs are longer
than some arbitrary number of bytes. In our case, the threshold seems to
be 4096 bytes, which happens to be the length of the url generated for
an eosnonfree eos3.5 id image build.

Move the payload to the POST request body, which does not have a (low)
limit.

https://phabricator.endlessm.com/T24918