needles: Ignore Plymouth boot animation
If the Plymouth boot animation is cut off before it finishes, due to the
boot process going faster than expected, we will never actually match
its final frame. Perversely, this would mean the tests would be stuck
expecting the system to boot, while the system would have actually
booted ages ago.

Fix that by ignoring the animation area on the screen, so we essentially
match the background colour and logo. Waiting for the animation to
complete is then the responsibility of the next test. This is typically
_fbe_install.pm, which waits a long time for the first screen of the FBE
to appear.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T24969