Run & fix fbe_language_switch on ISOs
Previously, the test was run against the (not-installed) ISO image,
which kicks the FBE into its live-image mode. This is in some sense
good, because it's nice to test that page of the FBE has its
translations. However, it didn't have any logic to check the LIVE flag;
and that flag wasn't set anyway.

To fix this, teach fbe_language_switch.pm to switch on whether LIVE is
set, just like the normal _fbe.pm test does; then adjust the test suite
definition to run it with LIVE set against ISOs, and with LIVE unset
against raw disk images.

https://phabricator.endlessm.com/T25144