tests: Update Travel test to almost pass
Unfortunately, it’s not currently possible to get it to completely pass
(see https://phabricator.endlessm.com/T25243), but this gets the test a
lot closer. Add various exclusion areas to needles to ignore content
which changes regularly, and reduce the test’s dependence on using
hard-coded articles which it assumed would appear at certain places in
the article lists. (Since the article lists change regularly, this isn’t
guaranteed.)

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T24855