needles: Ignore more variable elements of gnome-software main screen
When logged in as a different user, various additional elements can vary
in a way which breaks the test, but is irrelevant. If the number of apps
listed for that user is small, there will be no scroll bar and there
will be a ‘bottom of the scroll area’ indicator.

The USB category comes and goes seemingly randomly; see T24527.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T24005