tests: Disable Google Chrome installation by default
It gets in the way of a lot of tests, either by popping up polkit
authentication dialogues on the desktop; or by sending commands to the
first gnome-software instance which a test runs.

We don’t need Chrome for the tests. The desktop_chrome test should
continue to work because installing Chrome when the user explicitly
launches it should continue to work, even if automatic installation is
disabled.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T24005