desktop_chrome: assert that runtime matches LibreOffice's
All Endless OS systems have LibreOffice installed (unless users go out
of their way to remove it) so its runtime is extremely likely to be
installed. Chrome should use the same runtime, to minimize the chance
that installing it (which may happen automatically) will download a new
runtime.

LibreOffice has recently switched to org.freedesktop.Platform//18.08,
and we are in the process of switching Chrome to match. This test will
help us remember to do this again when LibreOffice next switches to a
newer runtime.

https://phabricator.endlessm.com/T25269