needles: Update LibreOffice Writer test string needles
As the toolbar layout has changed slightly in Writer, the test string
needles no longer match. The *_2 needle is from an older version of
Writer, and can be removed. The unsuffixed needle can just have an
exclusion area added for the toolbars to reduce the need for updates to
it in future when LO is updated again. We care about the test string
being rendered correctly here, rather than the pixel correctness of the
toolbars.

Signed-off-by: Philip Withnall <withnall@endlessm.com>