fbe_reformat: ignore most of image name
The name of the image to be installed includes the branch. As a result,
this test passes on master (since the screenshot was taken there) but
fails on eos3.5 (and soon, eos3.6). The size of the base image may also
vary.

Continue to match the "Endless OS" string at the start of the image
name, but exclude the remaining segment, including some space to the
right in case an image we want to test in future has a longer name.