WIP: desktop_shortcuts: adjust for new icon grid
There is now 1 more item in the bottom row of icons. As a result, all
the screenshots and coordinates need to be adjusted.

I got 80% of the way through this branch before realising that it would
perhaps be better to just clear the desktop at the start of this test so
that we don't have to repeat all this work every time we change the
default grid layout.

https://phabricator.endlessm.com/T25938