Update desktop screenshots
Now that the taskbar on free images is fixed to be App Center, Chrome,
Nautilus again, various screenshots stopped matching. Fix 'em all.

This is not enough to get desktop_background passing: it fails because,
after returning from a trip to VT3 (to reset the background), a polkit
prompt is shown:

> Authentication required
> Authentication is required to update information about software

I think this is a GNOME Software bug – I presume it is not calling
flatpak_installation_set_no_interaction() at some crucial juncture – and
plan to follow that up in due course.

https://phabricator.endlessm.com/T28610