Update Initial Setup welcome needles
In https://github.com/endlessm/eos-image-builder/commit/e06f2ee058 we
changed the languages to show at the top of the list. This also affects
the height of the rotating "Welcome!" "Bienvenue !" widget since this
takes on the height of the tallest translation of the string "Welcome!"
from that list of languages.

https://phabricator.endlessm.com/T26787