lib: Support username and set_password in user_console()
Add support for logging in to a VT as a specified user, and potentially
setting their password due to logging in for the first time.

Previously `user_console()` had only supported logging in as the `live`
user or the main user.

This will help the setup for parental controls, where
`parental_controls_setup` creates a new user account.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T28736